reafactor(ui): Move ModeToggle, PurlDetails to common
