+++ apps/clearance_ui/src/components/common/PurlDetails.tsx
