+++ apps/clearance_ui/src/components/package_library/columns.tsx
@@ -21,8 +21,8 @@ import {
     TooltipProvider,
     TooltipTrigger,
 } from "@/components/ui/tooltip";
+import PurlDetails from "@/components/common/PurlDetails";
 import DeletePackage from "@/components/delete_item/DeletePackage";
-import PurlDetails from "@/components/PurlDetails";
 
 // Get the table column datatype from the query response
 // Note: for reusing the component, this needs to be changed
